Remove `Ctl.Internal.Plutip` module
